requirements.txt
cerebras-cloud-sdk
flask
flask-cors
python-dotenv
requests
beautifulsoup4